fix: use 'exitable' timestamp as enqueue priority
This commit fixs the bug of using utxo_pos as priority. It should be the
exitable timestamp. See 'ExitableTimestamp' for timestamp logic.

Also, we change the timestamp to be using uint64 instead of uint256 to be able to pack
data into one field. The current standard timestamp is int32, but it would end at 2038,
which is not that far from now. Use 64 bits instead as most new 64 machines uses 64 bits.

-Year 2038 problem: https://en.wikipedia.org/wiki/Year_2038_problem
-GH issue: https://git.io/fj1we
-PR: https://github.com/omisego/plasma-contracts/pull/177

+++ plasma_framework/contracts/src/exits/utils/ExitableTimestamp.sol
@@ -15,13 +15,13 @@ library ExitableTimestamp {
     )
         internal
         pure
-        returns (uint256)
+        returns (uint64)
     {
         uint256 minExitableTimestamp = _now + _calculator.minExitPeriod;
 
         if (_isDeposit) {
-            return minExitableTimestamp;
+            return uint64(minExitableTimestamp);
         }
-        return Math.max(_blockTimestamp + (_calculator.minExitPeriod * 2), minExitableTimestamp);
+        return uint64(Math.max(_blockTimestamp + (_calculator.minExitPeriod * 2), minExitableTimestamp));
     }
 }
